# packages/contracts/http/auth/openapi.yaml
openapi: 3.1.0
info:
  title: Weav Identity Authentication API
  version: 1.5.0
  description: >-
    Identity-local JSON contract for registration, login, opaque refresh-token
    rotation, logout, current-user profile, self-service session management,
    password change, email verification, password recovery, and the bounded
    Google OAuth browser transport, including cookie-only web refresh/logout.
    The Gateway remains outside this milestone; the existing JSON/native
    refresh and logout contract is unchanged.
servers:
  - url: http://localhost:8081
    description: Local Identity service
...
security:
  - bearerAuth: []
paths:
...
  /auth/change-password:
    post:
      tags: [Authentication]
      operationId: changePassword
      summary: Change the authenticated user's local password
      description: >-
        Requires a valid access JWT and an ACTIVE user with a matching active
        session. Verifies the current local password, updates it, and revokes
        every session for the user, including the session authorizing this
        request, in one transaction. Wrong current credentials and accounts
        without a local password return the same generic 401 response.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: >-
            Password changed and all existing sessions revoked; the response
            has no body and the client must log in again.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: requestPasswordResetOtp
      summary: Request a password-reset OTP
      description: >-
        Public facade for POST /auth/otp/request with purpose PASSWORD_RESET.
        It uses the same account/IP budgets, bounded-admission path, generic
        receipt, and non-eligible-account behavior; it must not implement a
        separate recovery engine.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Generic password-recovery request receipt.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Replace a password using a verified reset grant
      description: >-
        Consumes the opaque reset grant at most once before the PostgreSQL
        mutation. The grant expires after 300 seconds, is bound to the
        PASSWORD_RESET purpose, user, and current credential fingerprint
        captured at request time, and is stored only as a hash. The database
        transaction rechecks ACTIVE status, the current stored email, and the
        current credential binding before it updates the
        password, records emailVerifiedAt when it is still null, and revokes
        all existing sessions. A database rollback after grant consumption
        does not restore the grant; the caller must request a new OTP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: >-
            Password replaced and all existing sessions revoked; the response
            has no body and is not cacheable.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /users/me/avatar:
    put:
      tags: [Users]
      operationId: replaceCurrentUserAvatar
      summary: Upload or replace the authenticated user's avatar
      description: >-
        Requires an ACTIVE user and matching active session. The server checks
        JPEG, PNG, or WebP magic bytes and decoder output, limits the source and
        normalized image to 2 MiB and 4096 by 4096 pixels, strips metadata by
        re-encoding, and stores under a server-owned user namespace. SVG,
        remote URLs, and client-selected object keys are rejected.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Updated public user profile.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      tags: [Users]
      operationId: getCurrentUserAvatar
      summary: Get a short-lived signed avatar URL
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signed URL and its absolute expiry.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUrlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [Users]
      operationId: deleteCurrentUserAvatar
      summary: Delete the authenticated user's avatar
      description: Clears the database reference before idempotent object cleanup.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Avatar reference removed; the response has no body.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        HS256 access JWT issued by Identity. Refresh tokens are opaque values
        supplied only in JSON request bodies and are never bearer credentials.
    internalServiceKey:
      type: apiKey
      in: header
      name: X-Internal-Service-Key
      description: V1 service-to-service key; never embed a real secret in this contract.
  headers:
    CacheControlNoStore:
      description: Prevents token-bearing responses from being cached.
      schema:
        type: string
        const: no-store
      example: no-store
...
    RetryAfter:
      description: Number of seconds until the caller should retry.
      schema:
        type: integer
        minimum: 1
      example: 60
    BearerChallenge:
      description: Bearer authentication challenge.
      schema:
        type: string
      example: Bearer
  responses:
    BadRequest:
      description: Request validation failed or the JSON body was malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Request validation failed
                  details:
                    - field: email
                      message: Email must be valid
                timestamp: '2026-09-05T08:30:00Z'
                status: 400
                path: /auth/register
            invalidProfile:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Request validation failed
                  details:
                    - field: displayName
                      message: Display name must be at most 120 characters
                timestamp: '2026-09-06T08:30:00Z'
                status: 400
                path: /users/me
            invalidPassword:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Request validation failed
                  details:
                    - field: newPassword
                      message: Password must be at most 72 bytes when encoded as UTF-8
                timestamp: '2026-09-06T08:30:00Z'
                status: 400
                path: /auth/change-password
    Unauthorized:
      description: Credentials, token, account, or session are invalid.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/BearerChallenge'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            unauthorized:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication failed
                  details: []
                timestamp: '2026-09-05T08:30:00Z'
                status: 401
                path: /auth/login
            invalidCurrentSession:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication failed
                  details: []
                timestamp: '2026-09-06T08:30:00Z'
                status: 401
                path: /users/me/sessions
            invalidCurrentPassword:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication failed
                  details: []
                timestamp: '2026-09-06T08:30:00Z'
                status: 401
                path: /auth/change-password
...
    NotFound:
      description: The target does not exist or is not owned by the current user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            sessionNotFound:
              value:
                error:
                  code: NOT_FOUND
                  message: Resource not found
                  details: []
                timestamp: '2026-09-06T08:30:00Z'
                status: 404
                path: /users/me/sessions/92d13ef9-3e27-4e2e-b788-7f1f5da5a5f7
...
    TooManyRequests:
      description: The applicable authentication rate limit was exceeded.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            rateLimited:
              value:
                error:
                  code: RATE_LIMITED
                  message: Too many requests
                  details: []
                timestamp: '2026-09-05T08:30:00Z'
                status: 429
                path: /auth/login
    ServiceUnavailable:
      description: >-
        A required OTP/recovery or avatar-storage dependency or bounded
        delivery admission is temporarily unavailable. The response is
        sanitized and does not reveal account existence, storage credentials,
        or provider details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            dependencyUnavailable:
              value:
                error:
                  code: DEPENDENCY_UNAVAILABLE
                  message: A required authentication dependency is temporarily unavailable
                  details: []
                timestamp: '2026-09-06T08:30:00Z'
                status: 503
                path: /auth/otp/request
    InternalError:
      description: Sanitized unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            internalError:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: An unexpected error occurred
                  details: []
                timestamp: '2026-09-05T08:30:00Z'
                status: 500
                path: /auth/login
  schemas:
...
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          $ref: '#/components/schemas/Password'
        newPassword:
          $ref: '#/components/schemas/Password'
...
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          $ref: '#/components/schemas/Email'
...
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [resetToken, newPassword]
      properties:
        resetToken:
          type: string
          minLength: 43
          maxLength: 43
          pattern: '^[A-Za-z0-9_-]{43}$'
          writeOnly: true
          description: Opaque at-most-once reset grant; send only in this JSON field.
        newPassword:
          $ref: '#/components/schemas/Password'
...
    AvatarUrlResponse:
      type: object
      additionalProperties: false
      required: [url, expiresAt]
      properties:
        url:
          type: string
          format: uri
          description: Short-lived signed URL; clients must not persist it.
        expiresAt:
          type: string
          format: date-time
          description: Absolute UTC expiry of the signed URL.
    OtpReceipt:
      type: object
      additionalProperties: false
      required: [challengeId, expiresIn, retryAfter]
      properties:
        challengeId:
          type: string
          minLength: 43
          maxLength: 43
          pattern: '^[A-Za-z0-9_-]{43}$'
          description: >-
            Opaque random identifier encoded as 43 unpadded base64url
            characters. It has the same response shape and identifier length
            whether or not a mail is sent.
        expiresIn:
          type: integer
          format: int64
          const: 300
          description: Challenge lifetime in seconds.
        retryAfter:
          type: integer
          format: int64
          const: 60
          description: Minimum resend cooldown in seconds.
...
    Email:
      type: string
      minLength: 3
      maxLength: 320
      pattern: "^ *[A-Za-z0-9.!#$%&'*+/=?^_`{|}~-]+@[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?(?:\\.[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?)+ *$"
      description: >-
        ASCII email identifier. Identity removes only boundary U+0020 SPACE
        characters and applies Locale.ROOT ASCII case folding for lookup and
        uniqueness. Other whitespace is rejected; no dot or plus rewriting is
        performed.
    Password:
      type: string
      format: password
      minLength: 8
      maxLength: 72
      writeOnly: true
      description: >-
        Passwords are neither trimmed nor normalized and must also be no more
        than 72 bytes when encoded as UTF-8 for BCrypt.
...
    UserResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - email
        - displayName
        - avatarStorageKey
        - systemRole
        - status
        - createdAt
        - updatedAt
        - emailVerifiedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          maxLength: 320
          description: Stored supported email spelling; never a password or token field.
        displayName:
          type: [string, 'null']
          maxLength: 120
        avatarStorageKey:
          type: [string, 'null']
          maxLength: 512
        systemRole:
          type: string
          enum: [USER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        createdAt:
          type: string
          format: date-time
          description: UTC RFC 3339 timestamp.
        updatedAt:
          type: string
          format: date-time
          description: UTC RFC 3339 timestamp.
        emailVerifiedAt:
          type: [string, 'null']
          format: date-time
          description: >-
            UTC RFC 3339 timestamp proving ownership of the current stored
            email, or null when it has not been verified.
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [error, timestamp, status, path]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        timestamp:
          type: string
          format: date-time
          description: UTC RFC 3339 timestamp.
        status:
          type: integer
          minimum: 400
          maximum: 599
        path:
          type: string
          maxLength: 2048
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          maxLength: 64
        message:
          type: string
          maxLength: 512
        details:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          maxLength: 256
        message:
          type: string
          maxLength: 512